# docs/openapi.yml
...
openapi: 3.1.0
components:
  schemas:
...
    Instant:
      type: string
      format: date-time
      examples:
      - 2022-03-10T16:15:50Z
...
    ListServicesResponse:
      type: object
      properties:
        services:
          type: array
          items:
            $ref: "#/components/schemas/PdServiceDto"
...
    PdServiceDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        branch:
          type: string
        availableOnEnv:
          type: boolean
        healthPath:
          type: string
        createdAt:
          $ref: "#/components/schemas/Instant"
        environmentIds:
          type: array
          items:
            type: string
...
    SoftwareReleasedEvent:
      type: object
      required:
      - repoId
      - version
      properties:
        runId:
          type: string
        repoId:
          type: string
          pattern: \S
        projectId:
          type: string
        repoName:
          type: string
        application:
          type: string
        version:
          type: string
          pattern: \S
        priority:
          type: string
...
    UpsertServiceRequest:
      type: object
      properties:
        branch:
          type: string
        availableOnEnv:
          type: boolean
        healthPath:
          type: string
        environmentIds:
          type: array
          items:
            type: string
  securitySchemes:
    SecurityScheme:
      type: openIdConnect
      openIdConnectUrl: http://dev-qits-idp:8080/idp/.well-known/openid-configuration
      description: Authentication
paths:
...
  /deployments/api/events/software-released:
    post:
      summary: Announce one released version of one application; it deploys on the
        worker
      operationId: softwareReleased
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SoftwareReleasedEvent"
        required: true
      responses:
        "202":
          description: Accepted; the deployment runs on the worker
        "400":
          description: No repoId or no version
        "401":
          description: Gate on and no machine token presented
        "403":
          description: Not Allowed
      tags:
      - Pd Event Controller
      security:
      - SecurityScheme:
        - qits:system
...
  /deployments/api/services:
    get:
      summary: "Every service, with the environments each is linked into"
      operationId: listServices
      responses:
        "200":
          description: The services
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListServicesResponse"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Pd Service Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:admin-agent
        - qits:agent
  /deployments/api/services/{name}:
    put:
      summary: "Register or update one service, replacing its environment links"
      operationId: upsertService
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertServiceRequest"
        required: true
      responses:
        "200":
          description: The updated service
          content:
            application/json:
              schema: {}
        "201":
          description: The newly registered service
        "400":
          description: Validation failed
        "404":
          description: An environmentId names no environment
        "401":
          description: Gate on and no machine token presented
        "403":
          description: Gate on and the token is for another service
      tags:
      - Pd Service Controller
      security:
      - SecurityScheme:
        - qits:system
    delete:
      summary: Remove a service and its links
      operationId: deleteService
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: Removed
        "404":
          description: No such service
        "401":
          description: Gate on and no machine token presented
        "403":
          description: Gate on and the token is for another service
      tags:
      - Pd Service Controller
      security:
      - SecurityScheme:
        - qits:system
info:
  title: qits-platform-deployments API
  version: 2026.1008.164327